requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
